Capstone_Solutions/requirement.txt
pymupdf
openai
torch
transformers
langchain-community
langchain-google-genai
google-generativeai
pandas
langchain
pypdf2
fastapi
uvicorn
pytesseract
pdfminer.six
python-docx
fitz
langchain_community
gradio
langchain_experimental
langchain_openai
tabulate